# openapi.yml
openapi: 3.1.1
info:
  title: Deck API
  description: Rustdeck HTTP API
  version: 1.0.0

tags:
  - name: Client
  - name: Config

paths:
  # Client
  /api/client/config:
    get:
      tags: [Client]
      summary: Get config
      description: Get deck dimensions config
      responses:
        200:
          description: success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeckDimensionsConfig"

  /api/client/buttons:
    get:
      tags: [Client]
      summary: Get buttons
      description: Get current button screen
      responses:
        200:
          description: success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RenderedDeckButton"

  /api/client/click/{y}/{x}:
    post:
      tags: [Client]
      summary: Post click
      description: Send a click event on button (y;x)
      parameters:
        - name: y
          in: path
          required: true
          schema:
            type: integer
        - name: x
          in: path
          required: true
          schema:
            type: integer

  /api/client/screen/{id}:
    post:
      tags: [Client]
      summary: Switch screen
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string

  /api/client/icon/{id}:
    get:
      tags: [Client]
      summary: Get icon
      description: Get icon file (raw) by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string

  # Config
  /api/config/config:
    patch:
      tags: [Config]
      summary: Update config
      description: Update deck dimensions config
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeckDimensionsConfig"

  /api/config/button/{y}/{x}:
    parameters:
      - name: y
        in: path
        required: true
        schema:
          type: integer
      - name: x
        in: path
        required: true
        schema:
          type: integer
    get:
      tags: [Config]
      summary: Get raw button
      description: Get raw button (template) by position (y;x)
      responses:
        200:
          description: success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RawDeckButton"
    patch:
      tags: [Config]
      summary: Update raw button
      description: Update raw button template
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RawDeckButton"
    delete:
      tags: [Config]
      summary: Delete raw button
      description: Delete raw button template

  /api/config/buttons/swap:
    patch:
      tags: [Config]
      summary: Swap raw buttons
      description: Swap raw button templates by position (y;x)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                a:
                  $ref: "#/components/schemas/Position"
                b:
                  $ref: "#/components/schemas/Position"

  /api/config/list/variables:
    get:
      tags: [Config]
      summary: List variables
      description: Get a list of all available variables
      responses:
        200:
          description: success
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string

  /api/config/list/actions:
    get:
      tags: [Config]
      summary: List actions
      description: Get a list of all available actions
      responses:
        200:
          description: success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PluginActionsData"

  /api/config/list/screens:
    get:
      tags: [Config]
      summary: List screens
      description: Get a list of all button screens
      responses:
        200:
          description: success
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string

components:
  schemas:
    Position:
      type: object
      properties:
        y:
          type: integer
        x:
          type: integer

    DeckDimensionsConfig:
      type: object
      properties:
        cols:
          type: integer
        rows:
          type: integer

    RawDeckButton:
      type: object
      properties:
        template:
          type: string
        icon:
          type: [string, "null"]
        style:
          $ref: "#/components/schemas/DeckButtonStyle"
        on_click_action:
          type: [string, "null"]

    RenderedDeckButton:
      type: object
      properties:
        content:
          type: string
        icon:
          type: [string, "null"]
        style:
          $ref: "#/components/schemas/DeckButtonStyle"
        position:
          type: object
          properties:
            y:
              type: integer
            x:
              type: integer

    DeckButtonStyle:
      type: object
      properties:
        text_size:
          type: integer
        text_align:
          type: string
          enum:
            - top
            - center
            - bottom

    DeckActionArgs:
      type: object
      properties:
        id:
          type: string
        description:
          type: string
        type:
          type: number

    DeckAction:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        args:
          type: array
          items:
            $ref: "#/components/schemas/DeckActionArgs"

    PluginActionsData:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        actions:
          type: array
          items:
            type: string